backend/requirements.txt
google-genai>=0.8.0
google-cloud-firestore>=2.14.0
google-adk>=0.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
httpx>=0.26.0
